VISION/requirements.txt
streamlit>=1.31
tensorflow>=2.13
pillow>=10.0
numpy>=1.23
